+++ src/test/java/org/thema/graphab/CLIToolsTest.java
@@ -128,6 +128,10 @@ public class CLIToolsTest {
         assertTrue(f.exists());
         assertEquals(4, DefaultFeature.loadFeatures(f).size());
         
+        cli.execute(new String[]{"--project", prjFile, "--corridor", "maxcost=3", "format=raster"});
+        f = new File(prjDir, "Jeulien1-corridor-3.0.tif");
+        assertTrue(f.exists());
+        
         thrown.expect(IllegalArgumentException.class);
         cli.execute(new String[]{"--project", prjFile, "--corridor"});
     }